input_handler/easyocr_pdf/requirements.txt
about-time==4.2.1
absl-py==2.1.0
accelerate==0.27.2
aiohttp==3.9.3
aiosignal==1.3.1
alive-progress==3.1.5
altair==5.2.0
annotated-types==0.6.0
anyio==4.3.0
appdirs==1.4.4
asgiref==3.7.2
astunparse==1.6.3
attrs==23.2.0
autograd==1.6.2
azure-ai-textanalytics==5.3.0
azure-common==1.1.28
azure-core==1.30.1
backoff==2.2.1
bcrypt==4.1.2
beautifulsoup4==4.12.3
bertopic==0.16.0
bitsandbytes==0.43.0
blinker==1.7.0
blis==0.7.11
boto3==1.34.66
botocore==1.34.66
bpemb==0.3.5
build==1.1.1
cachetools==5.3.3
catalogue==2.0.10
certifi==2024.2.2
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.3.2
chroma-hnswlib==0.7.3
chromadb==0.4.15
click==8.1.7
cloudpathlib==0.16.0
cma==3.2.2
coloredlogs==15.0.1
confection==0.1.4
conllu==4.5.3
contextlib2==21.6.0
contourpy==1.2.0
cryptography==42.0.5
ctransformers==0.2.27
cycler==0.12.1
cymem==2.0.8
Cython==0.29.37
dataclasses-json==0.6.4
datasets==2.18.0
deepdiff==6.7.1
Deprecated==1.2.14
dill==0.3.8
distro==1.9.0
dm-tree==0.1.8
dnspython==2.6.1
docker-pycreds==0.4.0
easyocr==1.7.1
emoji==2.10.1
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.7.1/en_core_web_sm-3.7.1-py3-none-any.whl#sha256=86cc141f63942d4b2c5fcee06630fd6f904788d2f0ab005cce45aadb8fb73889
et-xmlfile==1.1.0
faiss-cpu==1.8.0
fastapi==0.110.0
fastchat==0.1.0
filelock==3.13.1
filetype==1.2.0
FlagEmbedding==1.2.5
flair==0.13.1
flatbuffers==23.5.26
fonttools==4.49.0
frozenlist==1.4.1
fsspec==2024.2.0
ftfy==6.2.0
future==1.0.0
gast==0.5.4
gdown==5.1.0
gensim==4.3.2
gguf==0.6.0
gitdb==4.0.11
GitPython==3.1.42
google-auth==2.28.1
google-pasta==0.2.0
googleapis-common-protos==1.62.0
gpt4all==2.2.1.post1
grapheme==0.6.0
greenlet==3.0.3
grpcio==1.62.0
grpcio-tools==1.62.0
h11==0.14.0
h2==4.1.0
h5py==3.10.0
hdbscan==0.8.33
hpack==4.0.0
httpcore==1.0.4
httptools==0.6.1
httpx==0.27.0
huggingface-hub==0.21.3
humanfriendly==10.0
hyperframe==6.0.1
idna==3.6
imageio==2.34.0
importlib-metadata==6.11.0
importlib_resources==6.1.2
InstructorEmbedding==1.0.1
intel-extension-for-pytorch==2.2.0
intel-extension-for-transformers==1.3.2
isodate==0.6.1
Janome==0.5.0
Jinja2==3.1.3
jmespath==1.0.1
joblib==1.3.2
JPype1==1.5.0
jsonpatch==1.33
jsonpath-python==1.0.6
jsonpointer==2.4
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
jstyleson==0.0.2
keras==3.0.5
kiwisolver==1.4.5
kubernetes==29.0.0
langchain==0.1.11
langchain-community==0.0.28
langchain-core==0.1.32
langchain-openai==0.0.8
langchain-text-splitters==0.0.1
langchainhub==0.1.15
langcodes==3.3.0
langdetect==1.0.9
langid==1.1.6
langsmith==0.1.23
lazy_loader==0.3
libclang==16.0.6
llvmlite==0.42.0
lxml==5.1.0
Markdown==3.5.2
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.21.1
matplotlib==3.8.3
mdurl==0.1.2
ml-dtypes==0.3.2
mmh3==4.1.0
monotonic==1.6
more-itertools==10.2.0
mpld3==0.5.10
mpmath==1.3.0
multidict==6.0.5
multiprocess==0.70.16
murmurhash==1.0.10
mypy-extensions==1.0.0
namex==0.0.7
natsort==8.4.0
networkx==3.1
neural-compressor==2.4.1
neural-speed==0.3
ninja==1.10.2.4
nltk==3.8.1
nncf==2.9.0
numba==0.59.1
numpy==1.26.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.19.3
nvidia-nvjitlink-cu12==12.3.101
nvidia-nvtx-cu12==12.1.105
oauthlib==3.2.2
onnx==1.15.0
onnxruntime==1.17.1
openai==1.14.2
opencv-python-headless==4.9.0.80
openpyxl==3.1.2
opentelemetry-api==1.23.0
opentelemetry-exporter-otlp-proto-common==1.23.0
opentelemetry-exporter-otlp-proto-grpc==1.23.0
opentelemetry-instrumentation==0.44b0
opentelemetry-instrumentation-asgi==0.44b0
opentelemetry-instrumentation-fastapi==0.44b0
opentelemetry-proto==1.23.0
opentelemetry-sdk==1.23.0
opentelemetry-semantic-conventions==0.44b0
opentelemetry-util-http==0.44b0
openvino==2024.0.0
openvino-nightly==2024.1.0.dev20240307
openvino-telemetry==2023.2.1
openvino-tokenizers==2024.0.0.0
opt-einsum==3.3.0
optimum==1.17.1
optimum-intel==1.15.2
ordered-set==4.1.0
orjson==3.9.15
overrides==7.7.0
packaging==23.2
pandas==2.1.4
pdfminer.six==20231228
phonenumbers==8.13.32
pillow==10.2.0
plotly==5.20.0
portalocker==2.8.2
posthog==3.5.0
pptree==3.1
preshed==3.0.9
presidio-analyzer==2.2.353
presidio-anonymizer==2.2.353
prettytable==3.10.0
protobuf==4.25.3
psutil==5.9.8
pulsar-client==3.4.0
py-cpuinfo==9.0.0
pyarrow==15.0.0
pyarrow-hotfix==0.6
pyasn1==0.5.1
pyasn1-modules==0.3.0
pyclipper==1.3.0.post5
pycocotools==2.0.7
pycparser==2.21
pycryptodome==3.20.0
pydantic==2.6.3
pydantic_core==2.16.3
pydeck==0.8.1b0
pydot==2.0.0
Pygments==2.17.2
pymongo==4.6.2
pymoo==0.6.1.1
PyMuPDF==1.23.26
PyMuPDFb==1.23.22
pynndescent==0.5.11
pyparsing==3.1.2
pypdf==4.1.0
PyPDF2==3.0.1
PyPika==0.48.9
pyproject_hooks==1.0.0
PySocks==1.7.1
pytesseract==0.3.10
python-bidi==0.4.2
python-dateutil==2.9.0.post0
python-docx==1.1.0
python-dotenv==1.0.1
python-iso639==2024.2.7
python-magic==0.4.27
python-multipart==0.0.9
pytorch_revgrad==0.2.0
pytz==2024.1
PyYAML==6.0.1
qdrant-client==1.7.3
rank-bm25==0.2.2
rapidfuzz==3.6.2
referencing==0.33.0
regex==2023.12.25
requests==2.31.0
requests-file==2.0.0
requests-oauthlib==1.3.1
rich==13.7.1
rpds-py==0.18.0
rsa==4.9
s3transfer==0.10.1
safetensors==0.4.2
schema==0.7.5
scikit-image==0.22.0
scikit-learn==1.4.1.post1
scipy==1.12.0
segtok==1.5.11
semver==3.0.2
sentence-transformers==2.3.1
sentencepiece==0.1.99
sentry-sdk==1.42.0
seqeval==1.2.2
setproctitle==1.3.3
shapely==2.0.3
simpletransformers==0.70.0
six==1.16.0
smart-open==6.4.0
smmap==5.0.1
sniffio==1.3.1
soupsieve==2.5
spacy==3.7.4
spacy-huggingface-pipelines==0.0.4
spacy-legacy==3.0.12
spacy-loggers==1.0.5
spacy-stanza==1.0.4
SQLAlchemy==2.0.28
sqlitedict==2.1.0
srsly==2.4.8
stanza==1.6.1
starlette==0.36.3
streamlit==1.32.2
sympy==1.12
tabula-py==2.9.0
tabulate==0.9.0
tenacity==8.2.3
tensorboard==2.16.2
tensorboard-data-server==0.7.2
tensorboardX==2.6.2.2
tensorflow==2.16.1
tensorflow-io-gcs-filesystem==0.36.0
termcolor==2.4.0
texttable==1.7.0
tf_keras==2.16.0
thinc==8.2.3
threadpoolctl==3.3.0
tifffile==2024.2.12
tiktoken==0.6.0
tldextract==5.1.2
tokenizers==0.15.2
toml==0.10.2
toolz==0.12.1
torch==2.2.1
torchaudio==2.2.1+cpu
torchvision==0.17.1+cpu
tornado==6.4
tqdm==4.66.2
transformer-smaller-training-vocab==0.3.3
transformers==4.38.2
triton==2.2.0
typer==0.9.0
types-requests==2.31.0.20240218
typing-inspect==0.9.0
typing_extensions==4.10.0
tzdata==2024.1
umap-learn==0.5.5
unstructured==0.12.5
unstructured-client==0.21.1
urllib3==1.26.18
uvicorn==0.27.1
uvloop==0.19.0
wandb==0.16.4
wasabi==1.1.2
watchdog==4.0.0
watchfiles==0.21.0
wcwidth==0.2.13
weasel==0.3.4
websocket-client==1.7.0
websockets==12.0
Werkzeug==3.0.1
Wikipedia-API==0.6.0
wrapt==1.16.0
xxhash==3.4.1
yacs==0.1.8
yarl==1.9.4
youtube-transcript-api==0.6.2
zipp==3.17.0
